backend/pyproject.toml
[project]
name = "manga-backend"
version = "0.1.0"
description = "Backend service for AI manga generation"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Spell Team" }
]

dependencies = [
  "fastapi~=0.111",
  "uvicorn[standard]~=0.30",
  "pydantic-settings~=2.3",
  "SQLAlchemy~=2.0",
  "asyncpg~=0.29",
  "aiosqlite~=0.19",
  "alembic~=1.13",
  "google-cloud-tasks~=2.16",
  "google-cloud-storage~=2.16",
  "google-auth~=2.35",
  "google-cloud-aiplatform~=1.68",
  "httpx~=0.27",
  "python-dotenv~=1.0"
]

[project.optional-dependencies]
dev = [
  "ruff~=0.6",
  "mypy~=1.11"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
